Move wait_for spec helper to separate module
Make the helper reusable by multiple specs by moving it to a helper
module. Then include that module in specs that use it.
