+++ spec/lib/puma/appsignal_spec.rb
@@ -1,4 +1,6 @@
 RSpec.describe "Puma plugin" do
+  include WaitForHelper
+
   let(:probe) { MockProbe.new }
   before do
     module Puma
@@ -74,18 +76,5 @@ RSpec.describe "Puma plugin" do
       # Minutely probes started and called
       wait_for("enough probe calls") { probe.calls >= 2 }
     end
-  end
-
-  def wait_for(name)
-    max_wait = 5_000
-    i = 0
-    while i <= max_wait
-      break if yield
-      i += 1
-      sleep 0.001
-    end
-
-    return unless i == max_wait
-    raise "Waited 5 seconds for #{name} condition, but was not met."
   end
 end
